# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Doctor:
    description: Doctor information
    properties:
      created_at:
        description: '@Description The time when the record was created'
        example: "2023-01-01T00:00:00Z"
        type: string
      deleted_at:
        description: '@Description The time when the record was deleted (soft delete)'
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        description: '@Description The unique identifier for the record'
        example: 1
        type: integer
      specialty:
        description: '@Description The doctor''s specialty'
        example: Cardiology
        type: string
      updated_at:
        description: '@Description The time when the record was last updated'
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: '@Description The associated user'
      user_id:
        description: '@Description The ID of the associated user'
        example: 1
        type: integer
    type: object
  models.DoctorAvailability:
    description: Doctor availability information
    properties:
      created_at:
        description: '@Description The time when the record was created'
        example: "2023-01-01T00:00:00Z"
        type: string
      deleted_at:
        description: '@Description The time when the record was deleted (soft delete)'
        example: "2023-01-01T00:00:00Z"
        type: string
      doctor:
        allOf:
        - $ref: '#/definitions/models.Doctor'
        description: '@Description The associated doctor'
      doctor_id:
        description: '@Description The ID of the doctor'
        example: 1
        type: integer
      end_time:
        description: '@Description The end time of availability'
        example: "2023-01-01T17:00:00Z"
        type: string
      id:
        description: '@Description The unique identifier for the record'
        example: 1
        type: integer
      is_available:
        description: '@Description Whether the doctor is available during this time'
        example: true
        type: boolean
      start_time:
        description: '@Description The start time of availability'
        example: "2023-01-01T09:00:00Z"
        type: string
      updated_at:
        description: '@Description The time when the record was last updated'
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.HistoryEntry:
    description: History entry information
    properties:
      created_at:
        description: '@Description The time when the record was created'
        example: "2023-01-01T00:00:00Z"
        type: string
      date:
        description: '@Description The date of the entry'
        example: "2023-01-01T00:00:00Z"
        type: string
      deleted_at:
        description: '@Description The time when the record was deleted (soft delete)'
        example: "2023-01-01T00:00:00Z"
        type: string
      health_issues:
        description: '@Description The health issues recorded'
        example: Πονοκέφαλος και πυρετός
        type: string
      id:
        description: '@Description The unique identifier for the record'
        example: 1
        type: integer
      medical_history_id:
        description: '@Description The ID of the associated medical history'
        example: 1
        type: integer
      treatment:
        description: '@Description The treatment prescribed'
        example: Συνταγογράφηση παυσίπονων και ανάπαυση
        type: string
      updated_at:
        description: '@Description The time when the record was last updated'
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  models.User:
    description: User information
    properties:
      created_at:
        description: '@Description The time when the record was created'
        example: "2023-01-01T00:00:00Z"
        type: string
      deleted_at:
        description: '@Description The time when the record was deleted (soft delete)'
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        description: '@Description The user''s email address'
        example: georgios.papadopoulos@example.com
        type: string
      first_name:
        description: '@Description The user''s first name'
        example: Γεώργιος
        type: string
      id:
        description: '@Description The unique identifier for the record'
        example: 1
        type: integer
      last_name:
        description: '@Description The user''s last name'
        example: Παπαδόπουλος
        type: string
      role:
        description: '@Description The user''s role in the system'
        example: patient
        type: string
      updated_at:
        description: '@Description The time when the record was last updated'
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a clinic management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Clinic Management API
  version: "1.0"
paths:
...
  /doctors:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Doctor'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List all doctors
      tags:
      - Doctors
    post:
      consumes:
      - application/json
      parameters:
      - description: Doctor object
        in: body
        name: doctor
        required: true
        schema:
          $ref: '#/definitions/models.Doctor'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Doctor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new doctor
      tags:
      - Doctors
  /doctors/{id}:
    delete:
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a doctor
      tags:
      - Doctors
    get:
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Doctor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a doctor by ID
      tags:
      - Doctors
    put:
      consumes:
      - application/json
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated doctor object
        in: body
        name: doctor
        required: true
        schema:
          $ref: '#/definitions/models.Doctor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Doctor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a doctor
      tags:
      - Doctors
  /doctors/{id}/availabilities:
    get:
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DoctorAvailability'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get doctor availabilities
      tags:
      - Doctors
    post:
      consumes:
      - application/json
      parameters:
      - description: Doctor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Availability object
        in: body
        name: availability
        required: true
        schema:
          $ref: '#/definitions/models.DoctorAvailability'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DoctorAvailability'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create doctor availability
      tags:
      - Doctors
  /doctors/availabilities/{id}:
    delete:
      parameters:
      - description: Availability ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete doctor availability
      tags:
      - Doctors
  /history:
    post:
      consumes:
      - application/json
      description: Add a new entry to a patient's medical history
      parameters:
      - description: History entry
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/models.HistoryEntry'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.HistoryEntry'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Add history entry
      tags:
      - History
  /history/{medicalHistoryID}:
    get:
      description: Retrieve the complete medical history for a patient
      parameters:
      - description: Medical History ID
        in: path
        name: medicalHistoryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HistoryEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get patient history
      tags:
      - History
  /history/{medicalHistoryID}/latest:
    delete:
      description: Delete the most recent entry in a patient's medical history
      parameters:
      - description: Medical History ID
        in: path
        name: medicalHistoryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete latest history entry
      tags:
      - History
  /history/latest:
    put:
      consumes:
      - application/json
      description: Update the most recent entry in a patient's medical history
      parameters:
      - description: Updated history entry
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/models.HistoryEntry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HistoryEntry'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update latest history entry
      tags:
      - History
  /history/search:
    get:
      description: Search for specific entries in a patient's medical history
      parameters:
      - description: Medical History ID
        in: query
        name: medicalHistoryID
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - description: Health issue
        in: query
        name: healthIssue
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HistoryEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Search patient history
      tags:
      - History
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"